refactor: model MemberRole pulled out from projects namespace
