+++ modules/search-query/src/main/scala/io/renku/search/query/parse/QueryParser.scala
@@ -20,10 +20,9 @@ package io.renku.search.query.parse
 
 import cats.data.NonEmptyList
 import cats.parse.{Parser as P, Parser0 as P0}
-import io.renku.search.model.{EntityType, Keyword}
 import io.renku.search.model.projects.Visibility
+import io.renku.search.model.{EntityType, Keyword, MemberRole}
 import io.renku.search.query.*
-import io.renku.search.model.projects.MemberRole
 
 private[query] object QueryParser {
   private def candidates(names: Set[String]): Set[String] =
