+++ modules/search-solr-client/src/main/scala/io/renku/search/solr/documents/PartialEntityDocument.scala
@@ -18,11 +18,11 @@
 
 package io.renku.search.solr.documents
 
-import io.bullet.borer.NullOptions.given
 import io.bullet.borer.*
+import io.bullet.borer.NullOptions.given
 import io.bullet.borer.derivation.{MapBasedCodecs, key}
 import io.renku.search.model.projects.*
-import io.renku.search.model.{Id, Keyword, Name}
+import io.renku.search.model.{Id, Keyword, MemberRole, Name}
 import io.renku.search.solr.documents.Project as ProjectDocument
 import io.renku.search.solr.schema.EntityDocumentSchema.Fields as SolrField
 import io.renku.solr.client.{DocVersion, EncoderSupport}
